change to toggler controller since adams branch uses results_controller
